Replaced instances of uint32_t with uintptr_t to be compatible with Connection.

+++ lib/rule-creation/game-rule-engine/include/InGameUserManager.h
@@ -15,14 +15,14 @@ class InGameUserManager {
     ~InGameUserManager(){};
 
     // TODO: (GameInstance) Figure out when/how to call this class's functions.
-    void addNewUser(uint32_t userID, Role role, GameEnvironment::Environment userStates);
-    void deleteUser(uint32_t userID);
-    GameEnvironment::Environment getStatesOfUser(uint32_t userID);
-    std::map<uint32_t, std::pair<Role, GameEnvironment::Environment>> getAllUserStates();
-    void setStatesOfUser(uint32_t userID, Role role, GameEnvironment::Environment states);
+    void addNewUser(intptr_t userID, Role role, GameEnvironment::Environment userStates);
+    void deleteUser(intptr_t  userID);
+    GameEnvironment::Environment getStatesOfUser(intptr_t userID);
+    std::map<intptr_t , std::pair<Role, GameEnvironment::Environment>> getAllUserStates();
+    void setStatesOfUser(intptr_t  userID, Role role, GameEnvironment::Environment states);
 
     private:
-    std::map<uint32_t, std::pair<Role, GameEnvironment::Environment>> m_userStates;
+    std::map<intptr_t, std::pair<Role, GameEnvironment::Environment>> m_userStates;
 
 };
